spelling mistakes

+++ server/api/cart.js
@@ -68,16 +68,21 @@ router.put('/', async (req, res, next) => {
 
 router.put('/:orderId', async (req, res, next) => {
   try {
-    const order = await Order.findByPk(req.params.orderId)
-    await order.update({
-      complete: true
-    })
-    await Cart.destroy({
-      where: {
-        orderId: order.id
-      }
-    })
-    res.sendStatus(204)
+    if (req.user === undefined) {
+      req.session.cart = []
+      res.sendStatus(204)
+    } else {
+      const order = await Order.findByPk(req.params.orderId)
+      await order.update({
+        complete: true
+      })
+      await Cart.destroy({
+        where: {
+          orderId: order.id
+        }
+      })
+      res.sendStatus(204)
+    }
   } catch (err) {
     next(err)
   }
